chore(flags): organise hardware related classes in common folder
