+++ packages/boards/flag/src/hardware/Battery.h
